PreDeployv15, getting wildly close

+++ RocketElevatorsAPI/Controllers/BuildingController.cs
@@ -63,7 +63,7 @@ namespace RocketElevatorsAPI.Controllers
             List<Building> customerBuildings = new List<Building>();
             foreach(Building building in buildingsAll)
             {
-                if (building.Customer_Id == id)
+                if ((long)building.Customer_Id == id)
                 {
                     customerBuildings.Add(building);
                 }
